+++ src/main/file.js
@@ -1,4 +1,6 @@
-import {enableSave, createWindowTabWithFormattedData} from './utils'
+import {dialog as Dialog, BrowserWindow} from 'electron'
+import Fs from 'fs'
+import {enableSave, createWindowTabWithFormattedDataFile} from './utils'
 let path = require('path')
 const _ = require('lodash')
 function makeCustomFormat(separator, delimiter) {
@@ -15,17 +17,68 @@ function makeCustomFormat(separator, delimiter) {
   }
 }
 
-function openFile(format) {
-  Dialog.showOpenDialog({
-    filters: format.filters
-  }, function(filenames) {
-    readFile(filenames, format)
+function saveAsCustom() {
+  let currentWindow = BrowserWindow.getFocusedWindow()
+  let dialog = new BrowserWindow({width: 200, height: 400})
+  dialog.setMenu(null)
+  dialog.once('closed', function() {
+    ipc.removeAllListeners('formatSelected')
+    dialog = null
   })
+  ipc.once('formatSelected', function(event, data) {
+    dialog.close()
+    let format = makeCustomFormat(data.separator, data.delimiter)
+    saveFileAs(format, currentWindow)
+  })
+  dialog.loadURL(`http://localhost:9080/#/customformat`)
+}
+
+function saveFileAs(format, currentWindow) {
+  if (!currentWindow) {
+    currentWindow = BrowserWindow.getFocusedWindow()
+  }
+  Dialog.showSaveDialog({
+    filters: format.filters,
+    defaultPath: global.tab.activeTitle
+  }, function(filename) {
+    if (filename === undefined) {
+      // console.log('returning as no filename was entered...')
+      return
+    }
+    if (savedFilenameExists(filename)) {
+      Dialog.showMessageBox(currentWindow, {
+        type: 'warning',
+        // title is not displayed on screen on macOS
+        title: 'Data not saved',
+        message:
+`The data was not saved to the file.
+You selected a file name that is already used in this Data Package.
+To save the data, choose a unique file name.`
+      })
+      return
+    }
+    // enableSave()
+    currentWindow.webContents.send('saveData', format, filename)
+    currentWindow.format = format
+    currentWindow.webContents.send('saveDataSuccess')
+  })
+}
+
+function savedFilenameExists(filename) {
+  let threshold = global.tab.activeFilename === filename ? 1 : 0
+  let length = global.tab.filenames.length
+  let filtered = _.without(global.tab.filenames, filename)
+  return length - filtered.length > threshold
+}
+
+function saveFile() {
+  let currentWindow = BrowserWindow.getFocusedWindow()
+  currentWindow.webContents.send('saveData', currentWindow.format, global.tab.activeFilename)
 }
 
 function openCustom() {
   // var window = BrowserWindow.getFocusedWindow()
-  var dialog = new BrowserWindow({width: 200, height: 400})
+  let dialog = new BrowserWindow({width: 200, height: 400})
   dialog.setMenu(null)
   dialog.once('closed', function() {
     ipc.removeAllListeners('formatSelected')
@@ -39,75 +92,57 @@ function openCustom() {
   dialog.loadURL(`http://localhost:9080/#/customformat`)
 }
 
-function filenameExists(filename) {
-  let threshold = global.tab.activeFilename ? 1 : 0
-  let length = global.tab.filenames.length
-  let filtered = _.without(global.tab.filenames, filename)
-  return length - filtered.length > threshold
-}
-
-function saveFileAs(format, window) {
-  if (!window) {
-    window = BrowserWindow.getFocusedWindow()
-  }
-  Dialog.showSaveDialog({
-    filters: format.filters,
-    defaultPath: global.tab.activeTitle
+export function importDataPackage() {
+  Dialog.showOpenDialog({
+    filters: [
+      {
+        name: '*',
+        extensions: ['zip']
+      }
+    ],
+    properties: ['openFile']
   }, function(filename) {
     if (filename === undefined) {
-      // console.log('returning as no filename was entered...')
       return
     }
-    if (filenameExists(filename)) {
-      // console.log('returning as filename exists...')
-      Dialog.showMessageBox(window, {
-        type: 'warning',
-        // title is not displayed on screen on macOS
-        title: 'Data not saved',
-        message: 'The data was not saved to the file.\n\nYou selected a file name that is already used in this Data Package.\n\nTo save the data, choose a unique file name.'
-      })
-      return
-    }
-
-    // enableSave()
-    window.webContents.send('saveData', format, filename)
-    window.format = format
-    window.webContents.send('saveDataSuccess')
+    let window = BrowserWindow.getFocusedWindow()
+    window.webContents.send('importDataPackage', filename)
   })
 }
 
-function saveAsCustom() {
-  var window = BrowserWindow.getFocusedWindow()
-  var dialog = new BrowserWindow({width: 200, height: 400})
-  dialog.setMenu(null)
-  dialog.once('closed', function() {
-    ipc.removeAllListeners('formatSelected')
-    dialog = null
+function openFile(format) {
+  Dialog.showOpenDialog({
+    filters: format.filters
+  }, function(filenames) {
+    readFile(filenames, format)
   })
-  ipc.once('formatSelected', function(event, data) {
-    dialog.close()
-    var format = makeCustomFormat(data.separator, data.delimiter)
-    saveFileAs(format, window)
-  })
-  dialog.loadURL(`http://localhost:9080/#/customformat`)
 }
 
-function saveFile() {
-  var window = BrowserWindow.getFocusedWindow()
-  window.webContents.send('saveData', window.format)
+function openedFilenameExists(filename) {
+  return _.indexOf(global.tab.filenames, filename) > -1
 }
 
 function readFile(filenames, format) {
-  if (filenames === undefined) {
-
-  } else {
-    var filename = filenames[0]
+  if (filenames !== undefined) {
+    let filename = filenames[0]
+    if (openedFilenameExists(filename)) {
+      Dialog.showMessageBox(BrowserWindow.getFocusedWindow(), {
+        type: 'warning',
+        // title is not displayed on screen on macOS
+        title: 'File not opened',
+        message: `The file was not opened.
+You selected a file name that is already used in this Data Package.
+A file may only be opened once.`
+      })
+      return
+    }
     Fs.readFile(filename, 'utf-8', function(err, data) {
       if (err) {
         console.log(err.stack)
+      } else {
+        createWindowTabWithFormattedDataFile(data, format, filename)
+        // enableSave()
       }
-      createWindowTabWithFormattedData(data, format)
-      // enableSave()
     })
   }
 }