+++ src/main/index.js
@@ -2,8 +2,8 @@
  * Set `__static` path to static files in production
  * https://simulatedgreg.gitbooks.io/electron-vue/content/en/using-static-assets.html
  */
-import {dialog, app} from 'electron'
-import {quitOrSaveDialog} from './utils'
+import {dialog, app, Menu, BrowserWindow} from 'electron'
+import {quitOrSaveDialog, createWindowTab} from './utils'
 import {readFile} from './file.js'
 import {menu as template} from './menu'
 
@@ -19,27 +19,6 @@ global.tab = {
   activeFilename: '',
   filenames: []
 }
-global.electron = require('electron')
-
-// global.app = electron.app
-global.request = electron.request
-global.BrowserWindow = electron.BrowserWindow
-global.Menu = electron.Menu
-global.Dialog = electron.dialog
-
-global.Fs = require('fs')
-// global.XLSX = require('xlsx')
-global.ipc = require('electron').ipcMain
-
-global.utils = require('./utils')
-global.datapackage = require('./datapackage')
-global.github = require('./github')
-global.schema = require('./schema')
-global.excel = require('./excel')
-// global.fileActions = require('./file')
-global.tools = require('./tools')
-global.validate = require('./validate')
-global.help = require('./help')
 
 // var mainWindow = null
 function createWindow() {
@@ -50,7 +29,7 @@ function createWindow() {
   if (filename) {
     readFile([filename])
   } else {
-    global.utils.createWindowTab()
+    createWindowTab()
   }
 }
 
@@ -59,17 +38,17 @@ function createWindow() {
 // }
 
 const isSecondInstance = app.makeSingleInstance((commandLine, workingDirectory) => {
-// Someone tried to run a second instance, we should focus our window.
+  // Someone tried to run a second instance, we should focus our window.
   console.log('Attempted to open a second instance. Disallowing...')
   let firstWindow = BrowserWindow.getAllWindows()[0]
   if (firstWindow) {
-    if (firstWindow.isMinimized()) firstWindow.restore()
+    if (firstWindow.isMinimized()) { firstWindow.restore() }
     firstWindow.focus()
   }
 })
 
 if (isSecondInstance) {
-  console.log('second app detected. Quitting this.')
+  console.log('Data curator is already open. Quitting this application.')
   app.quit()
 }