+++ src/renderer/store/modules/hots.js
@@ -1,6 +1,7 @@
 const state = {
   hotTabs: {},
-  packageProperties: {}
+  packageProperties: {},
+  provenanceProperties: {}
 }
 
 function getHotColumnPropertiesFromPropertyObject(property) {
@@ -18,11 +19,11 @@ function getHotColumnPropertiesFromPropertyObject(property) {
 }
 
 const getters = {
+  getHotTabs: state => {
+    return state.hotTabs
+  },
   getAllHotColumnPropertiesFromHotId: (state, getters) => (hotId) => {
     return state.hotTabs[hotId].columnProperties || []
-  },
-  getHotTableSchema: (state, getters) => (hotId) => {
-    return state.hotTabs[hotId].tableSchema
   },
   getHotIdFromTabId: (state, getters) => (tabId) => {
     return new Promise((resolve, reject) => {
@@ -36,6 +37,9 @@ const getters = {
         resolve(hotId)
       }
     })
+  },
+  getProvenance: state => {
+    return state.provenanceProperties
   },
   getMissingValuesFromHot: (state, getters) => (hotId) => {
     if (state.hotTabs[hotId].tableProperties) {
@@ -71,6 +75,9 @@ const getters = {
 }
 
 const mutations = {
+  pushProvenance(state, value) {
+    _.set(state.provenanceProperties, 'markdown', value)
+  },
   pushHotTab(state, hotTab) {
     let hotId = hotTab.hotId
     if (!hotId) {
@@ -82,25 +89,27 @@ const mutations = {
   },
   pushColumnProperty(state, property) {
     _.set(state.hotTabs, `${property.hotId}.columnProperties[${property.columnIndex}].${property.key}`, property.value)
-    mutations.mergeCurrentColumnPropertiesOverTableSchema(state, property.hotId)
+    // mutations.mergeCurrentColumnPropertiesOverTableSchema(state, property.hotId)
   },
   pushTableProperty(state, property) {
     _.set(state.hotTabs, `${property.hotId}.tableProperties.${property.key}`, property.value)
     console.log(state.hotTabs)
   },
+  // TODO : tidy up these schema methods so that:
+  // - once we have a schema guess, always point the properties to the schema fields and then update these
   pushTableSchemaProperty(state, property) {
     mutations.pushTableProperty(state, property)
-    let hotTab = state.hotTabs[property.hotId]
-    if (hotTab.tableSchema) {
-      hotTab.tableSchema.schema[property.key] = property.value
-    }
+    // let hotTab = state.hotTabs[property.hotId]
+    // if (hotTab.tableSchema) {
+    //   hotTab.tableSchema.schema[property.key] = property.value
+    // }
   },
   pushTableSchemaDescriptorProperty(state, property) {
     mutations.pushTableProperty(state, property)
-    let hotTab = state.hotTabs[property.hotId]
-    if (hotTab.tableSchema) {
-      hotTab.tableSchema.schema.descriptor[property.key] = property.value
-    }
+    // let hotTab = state.hotTabs[property.hotId]
+    // if (hotTab.tableSchema) {
+    //   hotTab.tableSchema.schema.descriptor[property.key] = property.value
+    // }
   },
   pushPackageProperty(state, property) {
     _.set(state.packageProperties, property.key, property.value)
@@ -109,39 +118,68 @@ const mutations = {
   pushMissingValues(state, hotMissingValues) {
     let hotId = hotMissingValues.hotId
     _.set(state.hotTabs, `${hotId}.tableProperties.missingValues`, hotMissingValues.missingValues)
-    if (state.hotTabs[hotId].tableSchema) {
-      mutations.overwriteTableSchemaDescriptorProperty(state, hotId, 'missingValues')
-    }
+    // if (state.hotTabs[hotId].tableSchema) {
+    //   mutations.overwriteTableSchemaDescriptorProperty(state, hotId, 'missingValues')
+    // }
   },
+  // TODO : atm just chucking in entire table - probabaly only need schema and headers
   pushTableSchema(state, hotTable) {
     console.log('pushing table schema')
-    let hotId = hotTable.hotId
-    let tableSchema = _.set(state.hotTabs, `${hotId}.tableSchema`, hotTable.tableSchema)
-    let columnProperties = mutations.mergeTableSchemaOverCurrentColumnProperties(state, hotId)
-    return tableSchema && columnProperties
+    // let hotId = hotTable.hotId
+    // let tableSchema = _.set(state.hotTabs, `${hotId}.tableSchema`, hotTable.tableSchema)
+    let isColumnPropertiesMerged = mutations.mergeTableSchemaOverCurrentColumnProperties(state, hotTable)
+    console.log(state.hotTabs)
+    // return tableSchema && isColumnPropertiesMerged
+    return isColumnPropertiesMerged
   },
-  mergeTableSchemaOverCurrentColumnProperties(state, hotId) {
+
+  // TODO : once this behaviour is ensured (ie: we always overwrite) tidy up these 'merge' methods so that:
+  // - once we have a schema guess, always point the columnProperties to the schema fields and then update these
+  mergeTableSchemaOverCurrentColumnProperties(state, hotIdSchema) {
+    let hotId = hotIdSchema.hotId
     let hotTab = state.hotTabs[hotId]
-    let tableSchemaProperties = hotTab.tableSchema.schema.descriptor.fields
+    //    let tableSchemaProperties = hotTab.tableSchema.schema.descriptor.fields
+    // let schemaFields = hotIdSchema.schema.descriptor.fields
     if (!hotTab.columnProperties) {
       hotTab.columnProperties = []
     }
-    return _.merge(hotTab.columnProperties, tableSchemaProperties)
+    console.log('examining table schema descriptor...')
+    console.log(hotIdSchema.schema.descriptor)
+    // we cannot mutate the vuex state itself (in lodash call) - we can only assign a new value
+    let columnProperties = [...hotTab.columnProperties]
+    let isMerged = _.merge(columnProperties, hotIdSchema.schema.descriptor.fields)
+    state.hotTabs[hotId].columnProperties = columnProperties
+    return isMerged
   },
-  mergeCurrentColumnPropertiesOverTableSchema(state, hotId) {
-    let hotTab = state.hotTabs[hotId]
-    let tableSchema = hotTab.tableSchema
-    if (tableSchema) {
-      let tableSchemaProperties = hotTab.tableSchema.schema.descriptor.fields
-      let columnProperties = hotTab.columnProperties || []
-      _.merge(tableSchemaProperties, columnProperties)
-    }
-  },
+  // mergeTableSchemaOverCurrentColumnProperties(state, hotId) {
+  //   let hotTab = state.hotTabs[hotId]
+  //   let tableSchemaProperties = hotTab.tableSchema.schema.descriptor.fields
+  //   if (!hotTab.columnProperties) {
+  //     hotTab.columnProperties = []
+  //   }
+  //   // we cannot mutate the vuex state itself (in lodash call) - we can only assign a new value
+  //   let columnProperties = [...hotTab.columnProperties]
+  //   let isMerged = _.merge(columnProperties, tableSchemaProperties)
+  //   state.hotTabs[hotId].columnProperties = columnProperties
+  //   return isMerged
+  // },
+  // TODO : once this behaviour is ensured (ie: we always overwrite) tidy up these 'merge' methods so that:
+  // - once we have a schema guess, always point the columnProperties to the schema fields and then update these
+  // mergeCurrentColumnPropertiesOverTableSchema(state, hotId) {
+  //   let hotTab = state.hotTabs[hotId]
+  //   let tableSchema = hotTab.tableSchema
+  //   if (tableSchema) {
+  //     let tableSchemaProperties = [...hotTab.tableSchema.schema.descriptor.fields]
+  //     let columnProperties = hotTab.columnProperties || []
+  //     let isMerged = _.merge(tableSchemaProperties, columnProperties)
+  //     state.hotTabs[hotId].tableSchema.schema.descriptor.fields = tableSchemaProperties
+  //   }
+  // },
   destroyHotTab(state, hotId) {
     _.unset(state.hotTabs, hotId)
   },
-  destroyHotTabFromTabId(state, tabId) {
-    let hotId = getters.getHotIdFromTabId(tabId)
+  async destroyHotTabFromTabId(state, tabId) {
+    let hotId = await getters.getHotIdFromTabId(tabId)
     _.unset(state.hotTabs, hotId)
   },
   resetAllColumnPropertiesForHotId(state, hotId) {